chore(deps): bump nx from 15.9.4 to 16.1.0

+++ decorate-angular-cli.js
@@ -27,10 +27,10 @@ const cp = require('child_process');
 const isWindows = os.platform() === 'win32';
 let output;
 try {
-	output = require('@nrwl/workspace').output;
+	output = require('@nx/workspace').output;
 } catch (e) {
 	console.warn(
-		'Angular CLI could not be decorated to enable computation caching. Please ensure @nrwl/workspace is installed.'
+		'Angular CLI could not be decorated to enable computation caching. Please ensure @nx/workspace is installed.'
 	);
 	process.exit(0);
 }
@@ -49,11 +49,7 @@ function symlinkNgCLItoNxCLI() {
 			 * Such that it works in all shells and works with npx.
 			 */
 			['', '.cmd', '.ps1'].forEach((ext) => {
-				if (fs.existsSync(nxPath + ext))
-					fs.writeFileSync(
-						ngPath + ext,
-						fs.readFileSync(nxPath + ext)
-					);
+				if (fs.existsSync(nxPath + ext)) fs.writeFileSync(ngPath + ext, fs.readFileSync(nxPath + ext));
 			});
 		} else {
 			// If unix-based, symlink
@@ -61,9 +57,7 @@ function symlinkNgCLItoNxCLI() {
 		}
 	} catch (e) {
 		output.error({
-			title:
-				'Unable to create a symlink from the Angular CLI to the Nx CLI:' +
-				e.message
+			title: 'Unable to create a symlink from the Angular CLI to the Nx CLI:' + e.message
 		});
 		throw e;
 	}
